@@ -1,6 +1,6 @@
 package br.com.alura.aluraviagens.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class DiasUtil {
 
